+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version latest (0.18.0rc?, 2019-12-05?)
     - add `regions_set` and `regions_str` as well as service column
     - add date start/end
     - apply the same binning class to redshift data
+    - implement new display/email functions for the binned data
 
 
 Version 0.17.2 (2019-12-05)